requirements.txt
Django
python-dotenv
psycopg2-binary
pillow
